feat: 버튼 액션 메서드 addClient 구현, ClientStackView 내 addClient, setAutoLayout 구현
